+++ src/repl.rs
@@ -1,17 +1,5 @@
 // Copyright (C) 2023 Vince Vasta
 // SPDX-License-Identifier: Apache-2.0
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
 
 //! REPL for dply expressions.
 use anyhow::{anyhow, Result};
@@ -62,7 +50,7 @@ pub fn run() -> Result<()> {
     println!("Welcome to dply {}", env!("CARGO_PKG_VERSION"));
     println!("Use Tab for completions, arrows to move around, and Enter for selection.");
     println!("Enter twice with an empty line to execute the pipeline.");
-    println!("Read a file with 'parquet' or 'csv' to get columns completions.");
+    println!("Read a file with 'parquet', 'json', or 'csv' to get columns completions.");
     println!("For columns only completions start completions with a dot.");
 
     let prompt = DefaultPrompt {