+++ tests/functions/group_by.rs
@@ -1,21 +1,9 @@
 // Copyright (C) 2023 Vince Vasta
 // SPDX-License-Identifier: Apache-2.0
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
 use anyhow::Result;
 use indoc::indoc;
 
-use dply::interpreter;
+use super::assert_interpreter;
 
 #[test]
 fn group_by_mean_sd_var() -> Result<()> {
@@ -31,10 +19,9 @@ fn group_by_mean_sd_var() -> Result<()> {
             arrange(desc(n)) |
             show()
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (5, 5)
@@ -67,10 +54,9 @@ fn group_by_min_max() -> Result<()> {
             arrange(desc(n)) |
             show()
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (5, 4)
@@ -106,10 +92,9 @@ fn group_by_median_quantile() -> Result<()> {
             arrange(desc(n)) |
             show()
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (5, 7)
@@ -145,10 +130,9 @@ fn summarize_median_quantile() -> Result<()> {
             arrange(desc(n)) |
             show()
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (1, 6)
@@ -177,10 +161,9 @@ fn group_by_list() -> Result<()> {
             ) |
             show()
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (2, 3)
@@ -207,10 +190,9 @@ fn group_by_list() -> Result<()> {
             unnest(amounts, fares) |
             show()
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (13, 3)
@@ -251,10 +233,9 @@ fn summarize_list() -> Result<()> {
             ) |
             show()
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (1, 3)
@@ -279,10 +260,9 @@ fn summarize_list() -> Result<()> {
             unnest(amounts, fares) |
             show()
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (9, 3)