+++ tests/functions/relocate.rs
@@ -1,21 +1,9 @@
 // Copyright (C) 2023 Vince Vasta
 // SPDX-License-Identifier: Apache-2.0
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
 use anyhow::Result;
 use indoc::indoc;
 
-use dply::interpreter;
+use super::assert_interpreter;
 
 #[test]
 fn relocate_default() -> Result<()> {
@@ -24,10 +12,9 @@ fn relocate_default() -> Result<()> {
             relocate(payment_type, passenger_count) |
             head(1)
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (1, 19)
@@ -50,10 +37,9 @@ fn relocate_before_first() -> Result<()> {
             relocate(payment_type, passenger_count, before = VendorID) |
             head(1)
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (1, 19)
@@ -76,10 +62,9 @@ fn relocate_before() -> Result<()> {
             relocate(payment_type, passenger_count, before = fare_amount) |
             head(1)
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (1, 19)
@@ -102,10 +87,9 @@ fn relocate_after() -> Result<()> {
             relocate(payment_type, passenger_count, after = fare_amount) |
             head(1)
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (1, 19)
@@ -128,10 +112,9 @@ fn relocate_after_last() -> Result<()> {
             relocate(payment_type, passenger_count, after = airport_fee) |
             head(1)
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (1, 19)
@@ -154,10 +137,9 @@ fn relocate_same_col() -> Result<()> {
             relocate(payment_type, passenger_count, after = passenger_count) |
             head(1)
     "#};
-    let output = interpreter::eval_to_string(input)?;
 
-    assert_eq!(
-        output,
+    assert_interpreter!(
+        input,
         indoc!(
             r#"
             shape: (1, 19)